move useMap to hooks folder
